s19b: move jeep actor functions from jeep_mrl.c to jeep.c
The leading jeep_mrl.c functions operate on the jeep actor, whose
fields overlapped the passengers' m_segs1[17]. jeep.obj links right
before jeep_mrl.obj so the overlay layout is unchanged. Also type
the passenger rots/adjust vectors and light matrices.
